Edit index.html add recipes directory and add waffles.html inside it
Add the main header to the page index.html, as well as a link to
the waffles recipe, waffles.html.

Create a directory for recipes called "recipes", create the first
recipe page inside it called waffles.html.

Add the main header to the page waffles.html, a brief description and
an image of waffles, as well as ingredients using an unordered list
and steps using an ordered list.